requirements.txt
numpy
pandas
requests
PyYAML
langchain-community
langchain_ollama
faiss-cpu
langgraph
langfuse
flask
flask-cors
tqdm